rename configAndIndexingFunctions.ts
