+++ packages/core/src/sync/events.test.ts
@@ -6,7 +6,7 @@ import {
   getBlocksConfigAndIndexingFunctions,
   getErc20ConfigAndIndexingFunctions,
 } from "@/_test/utils.js";
-import { buildConfigAndIndexingFunctions } from "@/build/configAndIndexingFunctions.js";
+import { buildConfigAndIndexingFunctions } from "@/build/config.js";
 import type {
   BlockEvent,
   LogEvent,
